refactor: prefix receive modal with eth keyword
