+++ src/frontend/src/eth/components/receive/EthReceive.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { modalStore } from '$lib/stores/modal.store';
 	import { addressNotCertified } from '$lib/derived/address.derived';
-	import ReceiveModal from '$eth/components/receive/ReceiveModal.svelte';
+	import EthReceiveModal from '$eth/components/receive/EthReceiveModal.svelte';
 	import { metamaskNotInitialized } from '$eth/derived/metamask.derived';
 	import { waitWalletReady } from '$lib/services/actions.services';
 	import { modalReceive } from '$lib/derived/modal.derived';
@@ -29,5 +29,5 @@
 <ReceiveButton {compact} on:click={async () => await openReceive()} />
 
 {#if $modalReceive && $modalStore?.data === modalId}
-	<ReceiveModal />
+	<EthReceiveModal />
 {/if}
