+++ src/frontend/src/eth/components/receive/EthReceiveModal.svelte
@@ -4,7 +4,7 @@
 	import { Modal } from '@dfinity/gix-components';
 	import Copy from '$lib/components/ui/Copy.svelte';
 	import { networkAddress, networkEthereum } from '$lib/derived/network.derived';
-	import ReceiveMetamask from '$eth/components/receive/ReceiveMetamask.svelte';
+	import EthReceiveMetamask from '$eth/components/receive/EthReceiveMetamask.svelte';
 	import { i18n } from '$lib/stores/i18n.store';
 </script>
 
@@ -24,7 +24,7 @@
 		<ReceiveQRCode address={$networkAddress ?? ''} />
 
 		{#if $networkEthereum}
-			<ReceiveMetamask />
+			<EthReceiveMetamask />
 		{/if}
 	</div>
