+++ src/frontend/src/lib/components/tokens/TokenReceiveSend.svelte
@@ -2,7 +2,7 @@
 	import type { Token } from '$lib/types/token';
 	import { isNetworkIdICP } from '$lib/utils/network.utils';
 	import IcReceive from '$icp/components/receive/IcReceive.svelte';
-	import Receive from '$eth/components/receive/Receive.svelte';
+	import EthReceive from '$eth/components/receive/EthReceive.svelte';
 	import Send from '$eth/components/send/Send.svelte';
 	import IcSend from '$icp/components/send/IcSend.svelte';
 	import { runAndResetToken } from '$icp/services/token.services';
@@ -21,7 +21,7 @@
 		<IcReceive compact {token} on:nnsClose={close} />
 		<IcSend compact {token} on:nnsClose={close} />
 	{:else}
-		<Receive compact />
+		<EthReceive compact />
 		<Send compact {token} on:nnsClose={close} />
 	{/if}
 </div>
